refactor(viz): split files + autoplay + updated docs + added minimal tests
